feat(abstract-global-wallet-evm): add abstract global wallet connector (#32)
This PR adds a custom connector for Abstract's Global Wallet.

This wallet is an ERC-4337 compatible wallet based (contracts at
https://github.com/Abstract-Foundation/clave-contracts)

The wallet is managed by an embedded wallet signer secured by Privy's
social login infrastructure. Signing is accomplished via a popup
provided by the `@privy-io/cross-app-connect` package, which is a
dependency of this